feat(indicators): 新增数学变换类指标并完善文档体系
新增数学变换类指标 RECIP、INV_SQRT、LOG1P、EXPM1、DEG2RAD，完善 Rust 指标库至 103 个
添加指标场景速查表（四象限）和指标全量说明文档，优化教学路径
更新版本号至 0.1.70，补充相关测试用例和类型定义
