+++ docs/en/guide/indicator_scenario_quickref.md
@@ -1,0 +1,9 @@
+# Indicator Scenario Quickref (Quadrants)
+
+This page is currently maintained in Chinese first.
+
+- Chinese page: [按策略场景选指标速查表（四象限）](../../zh/guide/indicator_scenario_quickref.md)
+- Related docs:
+  - [AKQuant Indicator Reference](rust_indicator_reference.md)
+  - [Indicator Playbook](talib_indicator_playbook.md)
+  - [Chapter 16: AKQuant Indicator System and Practical Usage](../textbook/16_rust_indicators.md)
