+++ docs/en/guide/rust_indicator_reference.md
@@ -1,0 +1,10 @@
+# AKQuant Indicator Reference (All Supported Indicators)
+
+This page is currently maintained in Chinese first.
+
+- Chinese page: [AKQuant 指标全量说明（103 个）](../../zh/guide/rust_indicator_reference.md)
+- Related docs:
+  - [TA-Lib Top20 Plan](../advanced/talib_top20_plan.md)
+  - [Indicator Playbook](talib_indicator_playbook.md)
+  - [Indicator Scenario Quickref](indicator_scenario_quickref.md)
+  - [Chapter 16: AKQuant Indicators in Practice](../textbook/16_rust_indicators.md)
