+++ docs/en/textbook/16_rust_indicators.md
@@ -1,0 +1,10 @@
+# Chapter 16: AKQuant Indicator System and Practical Usage
+
+This chapter is currently maintained in Chinese first.
+
+- Chinese chapter: [第 16 章：AKQuant 指标全景与工程化使用](../../zh/textbook/16_rust_indicators.md)
+- Textbook home: [Chinese textbook index](../../zh/textbook/index.md)
+- Related docs:
+  - [AKQuant Indicator Reference](../guide/rust_indicator_reference.md)
+  - [Indicator Scenario Quickref](../guide/indicator_scenario_quickref.md)
+  - [Indicator Playbook](../guide/talib_indicator_playbook.md)
