+++ python/akquant/talib/__init__.py
@@ -29,13 +29,16 @@ from .funcs import (
     COSH,
     COVAR,
     CUBE,
+    DEG2RAD,
     DEMA,
     DIV,
     DX,
     EMA,
     EXP,
+    EXPM1,
     FLOOR,
     HT_TRENDLINE,
+    INV_SQRT,
     KAMA,
     LINEARREG,
     LINEARREG_ANGLE,
@@ -43,6 +46,7 @@ from .funcs import (
     LINEARREG_R2,
     LINEARREG_SLOPE,
     LN,
+    LOG1P,
     LOG10,
     MACD,
     MAMA,
@@ -68,6 +72,7 @@ from .funcs import (
     POW,
     PPO,
     RANGE,
+    RECIP,
     ROC,
     ROCP,
     ROCR,
@@ -114,6 +119,7 @@ __all__ = [
     "CORREL",
     "COVAR",
     "CUBE",
+    "DEG2RAD",
     "DIV",
     "LINEARREG",
     "LINEARREG_ANGLE",
@@ -133,6 +139,7 @@ __all__ = [
     "PPO",
     "POW",
     "PLUS_DI",
+    "EXPM1",
     "MINUS_DI",
     "MOD",
     "ULTOSC",
@@ -140,6 +147,7 @@ __all__ = [
     "T3",
     "MAMA",
     "LN",
+    "LOG1P",
     "MAX",
     "MAX2",
     "MAXINDEX",
@@ -172,10 +180,12 @@ __all__ = [
     "EXP",
     "ROUND",
     "CLIP",
+    "INV_SQRT",
     "SIGN",
     "SIN",
     "SINH",
     "SQ",
+    "RECIP",
     "MOM",
     "OBV",
     "WILLR",
