+++ src/indicators.rs
@@ -8,10 +8,10 @@ mod volume;
 
 pub use momentum::{CMO, MOM, ROC, ROCP, ROCR, ROCR100, RSI, WILLR};
 pub use moving_average::{
-    ABS, ACOS, ADD, APO, ASIN, ATAN, AVGDEV, CEIL, CLAMP01, CLIP, COS, COSH, CUBE, DEMA, DIV, EMA,
-    EXP, FLOOR, HT_TRENDLINE, KAMA, LN, LOG10, MACD, MAMA, MAX, MAX2, MAXINDEX, MIDPOINT, MIN,
-    MIN2, MININDEX, MINMAX, MINMAXINDEX, MOD, MULT, PPO, POW, RANGE, ROUND, SIGN, SIN, SINH, SMA,
-    SQ, SQRT, SUB, SUM, T3, TAN, TANH, TEMA, TRIMA, TRIX, WMA,
+    ABS, ACOS, ADD, APO, ASIN, ATAN, AVGDEV, CEIL, CLAMP01, CLIP, COS, COSH, CUBE, DEG2RAD, DEMA,
+    DIV, EMA, EXP, EXPM1, FLOOR, HT_TRENDLINE, INV_SQRT, KAMA, LN, LOG1P, LOG10, MACD, MAMA, MAX,
+    MAX2, MAXINDEX, MIDPOINT, MIN, MIN2, MININDEX, MINMAX, MINMAXINDEX, MOD, MULT, POW, PPO, RANGE,
+    RECIP, ROUND, SIGN, SIN, SINH, SMA, SQ, SQRT, SUB, SUM, T3, TAN, TANH, TEMA, TRIMA, TRIX, WMA,
 };
 pub use trend::{
     ADX, ADXR, AROON, AROONOSC, BETA, CCI, CORREL, COVAR, DX, LINEARREG, LINEARREG_ANGLE,
